refactor: deduplicate transaction query builder constructors with macros
The select/select_distinct/insert/update/delete/with methods were
copy-pasted identically across 5 transaction files (~650 lines).
Extract into sqlite_transaction_constructors! and
postgres_transaction_constructors! macros, mirroring the existing
postgres_builder_constructors! pattern.
