+++ src/transaction/sqlite/turso/mod.rs
@@ -148,149 +148,7 @@ impl<'conn, Schema> Transaction<'conn, Schema> {
         }
     }
 
-    /// Creates a SELECT query builder within the transaction
-    #[cfg(feature = "sqlite")]
-    pub fn select<'a, 'b, T>(
-        &'a self,
-        query: T,
-    ) -> TransactionBuilder<
-        'a,
-        'a,
-        Schema,
-        SelectBuilder<'b, Schema, SelectInitial, (), T::Marker>,
-        SelectInitial,
-    >
-    where
-        T: ToSQL<'b, SQLiteValue<'b>> + drizzle_core::IntoSelectTarget,
-    {
-        use drizzle_sqlite::builder::QueryBuilder;
-
-        let builder = QueryBuilder::new::<Schema>().select(query);
-
-        TransactionBuilder {
-            transaction: self,
-            builder,
-            _phantom: PhantomData,
-        }
-    }
-
-    /// Creates an INSERT query builder within the transaction
-    #[cfg(feature = "sqlite")]
-    pub fn insert<'a, Table>(
-        &'a self,
-        table: Table,
-    ) -> TransactionBuilder<
-        'a,
-        'a,
-        Schema,
-        InsertBuilder<'a, Schema, InsertInitial, Table>,
-        InsertInitial,
-    >
-    where
-        Table: SQLiteTable<'a>,
-    {
-        let builder = QueryBuilder::new::<Schema>().insert(table);
-        TransactionBuilder {
-            transaction: self,
-            builder,
-            _phantom: PhantomData,
-        }
-    }
-
-    /// Creates an UPDATE query builder within the transaction
-    #[cfg(feature = "sqlite")]
-    pub fn update<'a, Table>(
-        &'a self,
-        table: Table,
-    ) -> TransactionBuilder<
-        'a,
-        'a,
-        Schema,
-        UpdateBuilder<'a, Schema, UpdateInitial, Table>,
-        UpdateInitial,
-    >
-    where
-        Table: SQLiteTable<'a>,
-    {
-        let builder = QueryBuilder::new::<Schema>().update(table);
-        TransactionBuilder {
-            transaction: self,
-            builder,
-            _phantom: PhantomData,
-        }
-    }
-
-    /// Creates a SELECT DISTINCT query builder within the transaction
-    #[cfg(feature = "sqlite")]
-    pub fn select_distinct<'a, 'b, T>(
-        &'a self,
-        query: T,
-    ) -> TransactionBuilder<
-        'a,
-        'a,
-        Schema,
-        SelectBuilder<'b, Schema, SelectInitial, (), T::Marker>,
-        SelectInitial,
-    >
-    where
-        T: ToSQL<'b, SQLiteValue<'b>> + drizzle_core::IntoSelectTarget,
-    {
-        use drizzle_sqlite::builder::QueryBuilder;
-
-        let builder = QueryBuilder::new::<Schema>().select_distinct(query);
-
-        TransactionBuilder {
-            transaction: self,
-            builder,
-            _phantom: PhantomData,
-        }
-    }
-
-    /// Creates a DELETE query builder within the transaction
-    #[cfg(feature = "sqlite")]
-    pub fn delete<'a, T>(
-        &'a self,
-        table: T,
-    ) -> TransactionBuilder<
-        'a,
-        'a,
-        Schema,
-        DeleteBuilder<'a, Schema, DeleteInitial, T>,
-        DeleteInitial,
-    >
-    where
-        T: SQLiteTable<'a>,
-    {
-        let builder = QueryBuilder::new::<Schema>().delete(table);
-        TransactionBuilder {
-            transaction: self,
-            builder,
-            _phantom: PhantomData,
-        }
-    }
-
-    /// Creates a query with CTE (Common Table Expression) within the transaction
-    #[cfg(feature = "sqlite")]
-    pub fn with<'a, C>(
-        &'a self,
-        cte: C,
-    ) -> TransactionBuilder<
-        'a,
-        'a,
-        Schema,
-        QueryBuilder<'a, Schema, builder::CTEInit>,
-        builder::CTEInit,
-    >
-    where
-        C: builder::CTEDefinition<'a>,
-    {
-        let builder = QueryBuilder::new::<Schema>().with(cte);
-        TransactionBuilder {
-            transaction: self,
-            builder,
-            _phantom: PhantomData,
-        }
-    }
+    sqlite_transaction_constructors!('conn);
 
     /// Executes a raw query within the transaction
     pub async fn execute<'a, T>(&'a self, query: T) -> Result<u64, DrizzleError>
